requirements.txt
aws-cdk.aws-s3==1.71.0
aws-cdk.core==1.71.0
pyyaml
